Add error logic
