+++ packages/strapi-admin/admin/src/components/Users/ModalCreateBody/utils/schema.js
@@ -1,5 +1,6 @@
 import * as yup from 'yup';
 import { translatedErrors } from 'strapi-helper-plugin';
+// TODO update schema with utils
 
 const schema = yup.object().shape({
   firstname: yup.string().required(translatedErrors.required),
