+++ packages/strapi-admin/admin/src/validations/users/edit.js
@@ -1,0 +1,11 @@
+import * as yup from 'yup';
+import profileValidation from './profile';
+import rolesValidation from './roles';
+
+const schema = yup.object().shape({
+  ...profileValidation,
+  isActive: yup.bool(),
+  ...rolesValidation,
+});
+
+export default schema;
