+++ packages/strapi-admin/admin/src/validations/users/index.js
@@ -1,0 +1,3 @@
+export { default as editValidation } from './edit';
+export { default as profileValidation } from './profile';
+export { default as rolesValidation } from './roles';
