+++ packages/strapi-admin/admin/src/validations/users/roles.js
@@ -1,0 +1,8 @@
+import * as yup from 'yup';
+import { translatedErrors } from 'strapi-helper-plugin';
+
+const schema = {
+  roles: yup.array().required(translatedErrors.required),
+};
+
+export default schema;
